requirements.txt
SpeechRecognition
edge-tts
customtkinter
pygame
PyAudio
requests
wikipedia
pyautogui
python-dotenv
psutil
pyperclip
screen-brightness-control
pywhatkit
pytesseract
Pillow
duckduckgo-search
beautifulsoup4
lxml
pygetwindow
